Create code
Let's figure out how to determine which factors are most strongly associated with differences in treatment effects of an experiment in Africa which increases people's access to microcredit. We will look at change in profit from baseline to post experiment using a causal forest. Causal forests are similar to random forests, a powerful ML algorithm, except they are designed specifically to find splits in variables that emphasize differences in treatment effects. We will find the top predictors
without having to look at all 17 factors individually. An output of this analysis will also allow us to use the causal tree trained on this experiment dataset to be 
used to find future potential applicants of the program and score them on likelihood of positive impact. Various visualizations including a decision tree are used
to illustrate the complex relationship between people's factors and the treatment effect.